+++ relay-server/src/metrics_extraction/event.rs
@@ -1,5 +1,5 @@
 use relay_common::time::UnixTimestamp;
-use relay_dynamic_config::MetricExtractionConfig;
+use relay_dynamic_config::CombinedMetricExtractionConfig;
 use relay_event_schema::protocol::{Event, Span};
 use relay_metrics::Bucket;
 use relay_quotas::DataCategory;
@@ -48,7 +48,7 @@ impl Extractable for Span {
 pub fn extract_metrics(
     event: &Event,
     spans_extracted: bool,
-    config: &MetricExtractionConfig,
+    config: &CombinedMetricExtractionConfig<'_>,
     max_tag_value_size: usize,
     span_extraction_sample_rate: Option<f32>,
 ) -> Vec<Bucket> {
@@ -65,7 +65,7 @@ pub fn extract_metrics(
 
 fn extract_span_metrics_for_event(
     event: &Event,
-    config: &MetricExtractionConfig,
+    config: &CombinedMetricExtractionConfig<'_>,
     max_tag_value_size: usize,
     output: &mut Vec<Bucket>,
 ) {
@@ -1139,7 +1139,8 @@ mod tests {
         project.sanitize();
 
         let config = project.metric_extraction.ok().unwrap();
-        let metrics = extract_metrics(event.value().unwrap(), false, &config, 200, None);
+        let combined_config = (&config).into();
+        let metrics = extract_metrics(event.value().unwrap(), false, &combined_config, 200, None);
         insta::assert_debug_snapshot!(metrics);
     }
 
@@ -1310,7 +1311,8 @@ mod tests {
         project.sanitize();
 
         let config = project.metric_extraction.ok().unwrap();
-        let metrics = extract_metrics(event.value().unwrap(), false, &config, 200, None);
+        let combined_config = (&config).into();
+        let metrics = extract_metrics(event.value().unwrap(), false, &combined_config, 200, None);
         insta::assert_debug_snapshot!((&event.value().unwrap().spans, metrics));
     }
 
@@ -1371,7 +1373,8 @@ mod tests {
         project.sanitize();
 
         let config = project.metric_extraction.ok().unwrap();
-        let metrics = extract_metrics(event.value().unwrap(), false, &config, 200, None);
+        let combined_config = (&config).into();
+        let metrics = extract_metrics(event.value().unwrap(), false, &combined_config, 200, None);
 
         // When transaction.op:ui.load and mobile:true, HTTP spans still get both
         // exclusive_time metrics:
@@ -1407,7 +1410,8 @@ mod tests {
         project.sanitize();
 
         let config = project.metric_extraction.ok().unwrap();
-        let metrics = extract_metrics(event.value().unwrap(), false, &config, 200, None);
+        let combined_config = (&config).into();
+        let metrics = extract_metrics(event.value().unwrap(), false, &combined_config, 200, None);
 
         let usage_metrics = metrics
             .into_iter()
@@ -1431,7 +1435,10 @@ mod tests {
         config.sanitize(); // apply defaults for span extraction
 
         let extraction_config = config.metric_extraction.ok().unwrap();
-        generic::extract_metrics(span, &extraction_config)
+        generic::extract_metrics(
+            span,
+            &CombinedMetricExtractionConfig::from(&extraction_config),
+        )
     }
 
     /// Helper function for span metric extraction tests.
@@ -1633,7 +1640,8 @@ mod tests {
         project.sanitize();
 
         let config = project.metric_extraction.ok().unwrap();
-        let metrics = extract_metrics(event.value().unwrap(), false, &config, 200, None);
+        let combined_config = (&config).into();
+        let metrics = extract_metrics(event.value().unwrap(), false, &combined_config, 200, None);
 
         assert_eq!(metrics.len(), 4);